optimizing

+++ cleaner/sourceteq/source/m/gun/mgunwaves.h
@@ -4,7 +4,7 @@
 
 @interface mgunwaves:NSObject
 
--(void)restart:(CGFloat)initx inity:(CGFloat)inity centerx:(CGFloat)centerx centery:(CGFloat)centery finalx:(CGFloat)finalx finaly:(CGFloat)finaly pointermargin:(NSInteger)pointermargin;
+-(void)restart:(NSInteger)initx inity:(NSInteger)inity centerx:(NSInteger)centerx centery:(NSInteger)centery finalx:(NSInteger)finalx finaly:(NSInteger)finaly pointermargin:(NSInteger)pointermargin;
 -(void)clear;
 
 @property(strong, nonatomic)NSMutableArray<ggunwaves*> *items;
